refactor: integrate with template
feat: fps_counter
