--- src/main_old.rs
@@ -1,9 +1,0 @@
-// Disable console on Windows for non-dev builds.
-#![cfg_attr(not(feature = "dev"), windows_subsystem = "windows")]
-
-use bevy::prelude::*;
-use bevy_gamejam5::AppPlugin;
-
-fn main() -> AppExit {
-    App::new().add_plugins(AppPlugin).run()
-}
